Moving `lithium\tests\integration\net\SocketTest` integration tests to unit tests.
