+++ tests/cases/net/socket/StreamTest.php
@@ -20,7 +20,10 @@ class StreamTest extends \lithium\test\Unit {
 		'host' => 'google.com',
 		'port' => 80,
 		'timeout' => 2,
-		'classes' => array('request' => 'lithium\net\http\Request')
+		'classes' => array(
+			'request' => 'lithium\net\http\Request',
+			'response' => 'lithium\net\http\Response'
+		)
 	);
 
 	public function setUp() {
@@ -156,6 +159,20 @@ EOD;
 		$this->assertInstanceOf('lithium\net\http\Response', $result);
 		$this->assertPattern("/^HTTP/", (string) $result);
 		$this->assertTrue($stream->eof());
+	}
+
+	public function testStreamAdapter() {
+		$socket = new Stream($this->_testConfig);
+		$this->assertNotEmpty($socket->open());
+		$response = $socket->send();
+		$this->assertInstanceOf('lithium\net\http\Response', $response);
+
+		$expected = 'google.com';
+		$result = $response->host;
+		$this->assertEqual($expected, $result);
+
+		$result = $response->body();
+		$this->assertPattern("/<title[^>]*>301 Moved<\/title>/im", (string) $result);
 	}
 }
